thêm khối lượng trong thêm sản phẩm

+++ src/main/java/cleanmeat/model/Unit.java
@@ -5,11 +5,11 @@ import java.time.LocalDate;
 public class Unit {
     private int id;
     private String name;
-    private double amount;
+    private int amount;
     private LocalDate created_at;
     private LocalDate updated_at;
 
-    public Unit(int id, String name, double amount, LocalDate created_at, LocalDate updated_at) {
+    public Unit(int id, String name, int amount, LocalDate created_at, LocalDate updated_at) {
         this.id = id;
         this.name = name;
         this.amount = amount;
@@ -17,7 +17,7 @@ public class Unit {
         this.updated_at = updated_at;
     }
 
-    public Unit(String name, double amount) {
+    public Unit(String name, int amount) {
         this.name = name;
         this.amount = amount;
     }
@@ -45,7 +45,7 @@ public class Unit {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
